BRS(Backend)/requirements.txt
Flask
google-cloud-vision
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
